Refactor meta unit test to meta folder
Meta test checked if all data directories exist; will
raise an error if they don't exist in a new repo which
builds drem from scratch (i.e. if they aren't in source control)
